backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
aiofiles==24.1.0
requests==2.32.3
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv
aiohttp==3.11.11
mutagen==1.46.0
lrclibapi==0.3.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
troi==2025.8.6.3
pykakasi==2.3.0
beets
pyacoustid
pylast
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.23.2